Make digestof and hash() return USize instead of U64 and add hash64()
Default hash values will now match the platform machine word width for
performance. `hash64()` can be used if a low collision rate is needed.

+++ src/libponyc/ast/stringtab.c
@@ -24,7 +24,7 @@ typedef struct stringtab_entry_t
 
 static size_t stringtab_hash(stringtab_entry_t* a)
 {
-  return (size_t)ponyint_hash_block(a->str, a->len);
+  return ponyint_hash_block(a->str, a->len);
 }
 
 static bool stringtab_cmp(stringtab_entry_t* a, stringtab_entry_t* b)
